[#71] refactor:히스토리 페이지 메뉴 바 컴포넌트 분리
