+++ src/components/History/History/HistoryContent.js
@@ -1,23 +1,16 @@
-import React, { useEffect, useLayoutEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useInView } from 'react-intersection-observer';
-import HistoryCheckBox from './HisoryCheckBox';
 import * as S from './History.Style';
 import {
   getHistoryList,
-  deleteHistoryList,
-  moveToScrap,
   clearHistoryList,
-  undoHistoryList,
 } from 'store/history';
-import { deleteScrapItem } from 'store/scrap';
 import { useDispatch, useSelector } from 'react-redux';
 import PopUp from './HistoryPopup';
-import { SITE_LIST } from 'constant';
 import { useMediaQuery } from 'react-responsive';
 import theme from '../../../theme';
-import MobileMenuModal from './MobileMenuModal';
-import { MobileDeleteModal, MobileMoveScrapModal } from './MobilePopUpModal';
 import HistoryAlert from './HistoryAlert';
+import HistoryMenuBar from './HistoryMenuBar';
 const HistoryContent = ({isToggleOpen}) => {
   const { historyList, deleteHistoryResponse, readHistoryItemResponse, moveToScrapResponse, undoHistoryListResponse } =
     useSelector((state) => state.history);
@@ -30,70 +23,11 @@ const HistoryContent = ({isToggleOpen}) => {
   const [pageNum, setPageNum] = useState([1]);
   const [isPopOpen, setOpen] = useState(false);
   const [isMobileMenuOpen, setMobileMenu] = useState(false);
-  const [isMobileDeleteOpen, setMobileDeleteModal] = useState(false);
-  const [isMobileScrapOpen, setMobileScrapModal] = useState(false);
-  const [undoList, setUndoList] = useState([]);
   const [refAlert, inView] = useInView({
     threshold: 0.0,
     triggerOnce: false,
   });
   const isMobile = useMediaQuery({ query: `(max-width:${theme.deviceSizes.tabletL}` });
-  const showRead = () => {
-    if (command === 'read') {
-      setCommand(null);
-    } else {
-      setCommand('read');
-    }
-  };
-  const showNotRead = () => {
-    if (command === 'notRead') {
-      setCommand(null);
-    } else {
-      setCommand('notRead');
-    }
-  };
-  const moveToStorage = () => {
-    if (checkedList.length > 0) {
-      try {
-        checkedList.forEach((id) => {
-          dispatch(moveToScrap({ crawling_id: id }));
-        });
-        setUndoList(checkedList);
-        setCheckedList([]);
-      } catch (e) {
-        console.log(e);
-      } finally {
-        if (isMobile) {
-          setMobileScrapModal(true);
-        } else {
-          setOpen(true);
-        }
-      }
-    } else {
-      alert('이동할 메일을 선택해 주세요');
-    }
-  };
-  const deleteMail = () => {
-    if (checkedList.length > 0) {
-      try {
-        var deleteMailQuery = '';
-        checkedList.forEach((id) => {
-          deleteMailQuery += `notice-id=${id}&`;
-        });
-        dispatch(deleteHistoryList(deleteMailQuery));
-        setUndoList([checkedList]);
-        setCheckedList([]);
-      } catch (e) {
-        console.log(e);
-      } finally {
-        if (isMobile) {
-          setMobileDeleteModal(true);
-        }
-      }
-    } else {
-      alert('삭제할 메일을 선택해 주세요');
-    }
-  };
   const selectAllMail = (e) => {
     if (e.target.checked) {
       setCheckedList(
@@ -115,30 +49,8 @@ const HistoryContent = ({isToggleOpen}) => {
   const closePopUp = () => {
     setOpen(false);
   };
-  const openMobileMenu = () => {
-    setMobileMenu(!isMobileMenuOpen);
-  };
   const closeMobileMenu = () => {
     setMobileMenu(false);
-  };
-  const undoDelete = () => {
-    var undoMailQuery = '';
-    try {
-      undoList.forEach((id) => {
-        undoMailQuery += `notice-id=${id}&`;
-      });
-      dispatch(undoHistoryList(undoMailQuery));
-    } catch (e) {
-      console.log(e);
-    } finally {
-      setUndoList([]);
-      setList([]);
-      setMobileDeleteModal(false);
-    }
-  };
-  const undoMoveScrap = () => {
-    dispatch(deleteScrapItem(undoList));
-    setUndoList([]);
   };
   useEffect(() => {
     if (userInfo.isLoggedIn || deleteHistoryResponse || readHistoryItemResponse || moveToScrapResponse) {
@@ -186,77 +98,24 @@ const HistoryContent = ({isToggleOpen}) => {
     }
   }, [inView, readHistoryItemResponse, deleteHistoryResponse, undoHistoryListResponse, historyList]);
 
-  useEffect(() => {
-    if (isMobileDeleteOpen) {
-      setTimeout(() => {
-        setMobileDeleteModal(false);
-        setUndoList([]);
-      }, 4000);
-    } else if (isMobileScrapOpen) {
-      setTimeout(() => {
-        setMobileScrapModal(false);
-        setUndoList([]);
-      }, 4000);
-    }
-  }, [isMobileDeleteOpen, isMobileScrapOpen]);
   return (
     <>
       <S.PageWrapper onClick={isMobile ? closeMobileMenu : null}>
       {!isMobile && <PopUp isOpen={isPopOpen} closePopUp={closePopUp} />}
         <S.Content isOpen={isPopOpen}>
-          <S.MenuList>
-          <S.CheckBox>
-          <HistoryCheckBox
-                onClick={(e) => selectAllMail(e)}
-                checked={checkedList.length <= 0 || checkedList.length !== showList.length ? false : true}
-                readOnly
-              />
-            <S.SelectAll>
-              전체선택
-            </S.SelectAll>
-          </S.CheckBox>
-         
-            <S.MenuWrapper onClick={(e) => e.stopPropagation()}>
-              {!isMobile && (
-                <>
-                  <S.Menues onClick={() => showRead()} isClicked={command === 'read' ? true : false}>
-                    <S.MenuName>읽은 알림</S.MenuName>
-                  </S.Menues>
-                  <S.Menues onClick={() => showNotRead()} isClicked={command === 'notRead' ? true : false}>
-                    <S.MenuName>읽지 않은 알림</S.MenuName>
-                  </S.Menues>
-                </>
-              )}
-
-              <S.Menues onClick={() => moveToStorage()}>
-                <S.MenuLogo src="/asset/Storage.svg" />
-                <S.MenuName>{!isMobile ? '보관함으로이동' : '보관'}</S.MenuName>
-                <div onClick={(e) => e.stopPropagation()}>
-                  {isMobile && isMobileScrapOpen && (
-                    <MobileMoveScrapModal numberAlert={undoList.length} undo={undoMoveScrap} />
-                  )}
-                </div>
-              </S.Menues>
-              <S.Menues onClick={() => deleteMail()}>
-                <S.MenuLogo src="/asset/Delete.svg" />
-                <S.MenuName>삭제</S.MenuName>
-                <div onClick={(e) => e.stopPropagation()}>
-                  {isMobile && isMobileDeleteOpen && <MobileDeleteModal undo={undoDelete} />}
-                </div>
-              </S.Menues>
-              {isMobile && (
-                <>
-                  <S.MobileMenu src="/asset/MobileMenuDots.svg" onClick={openMobileMenu} />
-                  <MobileMenuModal
-                    isOpen={isMobileMenuOpen}
-                    showRead={showRead}
-                    showNotRead={showNotRead}
-                    command={command}
-                  />
-                </>
-              )}
-            </S.MenuWrapper>
-          </S.MenuList>
+          <HistoryMenuBar
+            selectAllMail={selectAllMail}
+            checkedList={checkedList}
+            showList={showList}
+            setCheckedList={setCheckedList}
+            setList={setList}
+            setOpen={setOpen}
+            isMobile={isMobile}
+            command={command}
+            setCommand={setCommand}
+            isMobileMenuOpen={isMobileMenuOpen}
+            setMobileMenu={setMobileMenu}
+          />
           <S.KeyWordAlertList>
             {showList?.map((mail) =>
               showList[showList.length - 1].id === mail.id ? (
